+++ src/QuickPick.js
@@ -1,6 +1,6 @@
 /* eslint-disable semi */
 const Core = require('./Core');
-const camelCase = require('lodash.camelcase');
+const camelCase = require('lodash/camelCase');
 const { showError } = require('./error-handling');
 
 module.exports = {